apply fixes

+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/controller/LinkExecution.scala
@@ -1,7 +1,7 @@
 package edu.uci.ics.amber.engine.architecture.controller
 
-class LinkExecution(totalReceiversCount: Int) extends Serializable {
-  private var currentCompletedCount = 0
+class LinkExecution(totalReceiversCount: Long) extends Serializable {
+  private var currentCompletedCount = 0L
 
   def incrementCompletedReceiversCount(): Unit = currentCompletedCount += 1
